requirements.txt
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
fastapi==0.61.2
h11==0.11.0
idna==2.10
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pydantic==1.7.2
pyparsing==2.4.7
pytest==6.1.2
python-dotenv==0.15.0
requests==2.25.0
six==1.15.0
starlette==0.13.6
toml==0.10.2
urllib3==1.26.2
uvicorn==0.12.2
